+++ npc.py
@@ -88,6 +88,12 @@ class Shopkeeper: #class for shopkeeper NPC's. Regular NPC's will use a similar 
                     f.header(response, 0.5)
                     for item in self.shop_inventory[response]:
                         print(f"{items_list[item].name}: {items_list[item].description}. [{f.limit(round(items_list[item].value+(items_list[item].value*(0.2-(self.relation*0.002)))), round(items_list[item].value*1.5), items_list[item].value)} Gold]", 0.2)
+                        if item in p.weapons_list:
+                            f.encyclopedia["Weapons"][item] = p.items_list[item]
+                        elif item in p.armor_list:
+                            f.encyclopedia["Armor"][item] = p.items_list[item]
+                        else:
+                            f.encyclopedia["Items"][item] = p.items_list[item]
                     f.header("", 0.5)
                     response2 = f.capitalize(input("Enter the name of the item you'd like to purchase, or 'Exit' to return to the main shop menu: "))
                     sleep(0.5)
@@ -144,6 +150,7 @@ def converse(npc): #called when the player visits an NPC. Plays an intro dialogu
     if npc.dialogue_track == -1:
         print(f"{npc.name}: {npc.intros['Meet']}", 3)
         npc.dialogue_track = 0
+        f.encyclopedia["Characters"][npc.name] = npc
     else:
         chosen_relation = -100
         for intro in npc.intros:
@@ -186,7 +193,7 @@ def converse(npc): #called when the player visits an NPC. Plays an intro dialogu
 Andre = Shopkeeper("Andre", "An old, grizzled blacksmith who's forged weapons and armor for many a naive adventurer.", 0, [], ["Chalgos"], [], ["Talk", "Quests", "Buy", "Sell", "Exit"], \
                    {"Meet": "Hmph... what are you, nother novice adventurer set on rushing to their deaths? Well if that's what you want, I may as well try to help you last as long as possible...", -100: "Bold of you to show your face around here... do your business quick and go.", -40: "Hm... What do you want...", -5: "Ah, you're back... honestly I'm somewhat surprised.", 20: f"Welcome back {p.player_name}, what can I do for ya?", 60: f"Well, if it isn't {p.player_name}! How can I be of service?"}, \
                    [["Bah, don't bother trying to convince me otherwise, I've seen plenty of your kind before... and they always end up the same way...", "But, I can't say I don't understand the desire to venture out into the great unknown and try to make a name for yourself.", "I considered it once... but fate had other plans in mind for me."], ["I'll try my best to equip you well enough that you at least have a chance out there.", "... Hopefully my wares will serve you better than it did the rest..."], ["I suppose I may as well try and impart some words of wisdom on you if you insist on this.", "If you want any hope of making it out there then you'd better listen close.", "First off, you'd best pay attention to the damage types of your weapons. Every weapon does different amounts of the various damage types, and enemies will be more resistant or weak to different ones.", "For example, slashing attacks won't do much against heavily armored foes, but there's little their armor can do against a solid blunt attack."]], \
-                   ["Mainhand", "Offhand", "Both", "Special", "Helmet", "Armor", "Leggings", "Gloves", "Boots"], [0, 1, 2], [], [])
+                   ["Mainhand", "Offhand", "Both", "Special", "Helmet", "Chest", "Leggings", "Gloves", "Boots"], [0, 1, 2], [], [])
 
 Malbras = Shopkeeper("Malbras", "The kind owner of the only tavern in Chalgos. He'll sell basic supplies and food necessary for adventuring.", 10, [], ["Chalgos"], [], ["Talk", "Quests", "Buy", "Sell", "Exit"], \
                     {"Meet": f"Hey there, you must be {p.player_name}, my name's Malbras, make yourself comfortable!", -100: "The only reason you can even show your face in here is because I can't afford a guard...", -60: "You're lucky I'm so lenient towards paying customers...", -10: "What'll it be today?", 30: f"Hey there {p.player_name}, what're ya lookin for today?", 60: f"Welcome in {p.player_name}! Make yourself right at home!"}, \