+++ statuses.py
@@ -18,6 +18,7 @@ player = c.player
 moves_list = c.moves_list
 specials_list = c.specials_list
 statuses_list = {} #Statuses are the full status effect with its own systems for application, recovery, and recurring impacts. They use effects, which are the back-end functions that actually do things.
+buffs_list = {} #Buffs are similar to statuses however they have unlimited duration and are tied to equipment. Buffs are gained by equipping equipment with buffs, and they are then lost when that piece of equipment is uneqipped.
 effects_list = {} #Empty dictionary of all effects waiting to be filled during initialization.
 operators = { #Used with the standard effects class for modularity purposes.
     '+': operator.add,
@@ -101,6 +102,46 @@ class Status: #Class for statuses, which are what control effects. Statuses do m
             target.statuses[index][1] -= 1
             self.duration = self.duration - 1 if target is player else self.duration
 
+class Buff: #Class for buffs, which are much more simplified than statuses as they can only affect the player, never have a duration, and will always either increase or decrease set values by the same amount.
+    def __init__(self, name, description, effect):
+        self.name = name
+        self.description = description
+        self.effect = effect
+        self.total = 0
+        statuses_list[self.name] = self
+        buffs_list[self.name] = self
+
+    def __str__(self):
+        operator = "+" if self.total >= 0 else ""
+        return f"{self.name}: {self.description} [{operator}{self.total}]"
+    
+    def __call__(*args):
+        pass
+    
+    def add(self, amount):
+        self.total += amount
+        if self.name not in player.statuses:
+            player.statuses.append([self.name, -1])
+        if "Affinity" in self.name:
+            effects_list["Affinity"](player, [self.effect, amount, "+"])
+        elif "Resistance" in self.name:
+            effects_list["Resistance"](player, [self.effect, amount, "+"])
+        else:
+            effects_list[self.effect](player, ["", amount, "+"])
+
+    def remove(self, amount):
+        self.total -= amount
+        if self.total == 0:
+            for index, status in enumerate(player.statuses):
+                if status[0] == self.name:
+                    del player.statuses[index]
+        if "Affinity" in self.name:
+            effects_list["Affinity"](player, [self.effect, amount, "-"])
+        elif "Resistance" in self.name:
+            effects_list["Resistance"](player, [self.effect, amount, "-"])
+        else:
+            effects_list[self.effect](player, ["", amount, "-"])
+
 class Standard_Effect: #Class for basic, formulaic effects while still being extremely modular.
     def __init__(self, name, message_type, verb, affect):
         self.name = name #name of the effect.
@@ -130,6 +171,11 @@ class Standard_Effect: #Class for basic, formulaic effects while still being ext
         p.effective_intelligence = operators[operator](p.effective_intelligence, amount) if self.affect == "Intelligence" else p.effective_intelligence
         p.effective_faith = operators[operator](p.effective_faith, amount) if self.affect == "Faith" else p.effective_faith
         p.speed = operators[operator](p.speed, amount) if self.affect == "Speed" else p.speed
+        p.xp_gain_multiplier = operators[operator](p.xp_gain_multiplier, amount) if self.affect == "XP Gain" else p.xp_gain_multiplier
+        if self.affect == "Affinity":
+            player.damage_affinities[effect[0]] += amount
+        if self.affect == "Resistance":
+            player.damage_resistances[effect[0]] += amount
 
 class Apply_Cure: #Simple class for cure effects, necessary for consumable items that apply a cure such as an antidote.
     def __init__(self, name):
@@ -154,9 +200,10 @@ def recur_statuses(target, exhaust=False): #command used to call every status th
 
 def cure_check(target, cure_all=False): #checks if any of the cure conditions are met for any of the targets statuses, and cures them if so.
     for index, ailment in enumerate(target.statuses):
-        for possible_cure in statuses_list[ailment[0]].cures:
-            if target.cures_list[possible_cure] == True or cure_all == True:
-                statuses_list[ailment[0]].cure(target, index)
+        if ailment[0] not in buffs_list:
+            for possible_cure in statuses_list[ailment[0]].cures:
+                if target.cures_list[possible_cure] == True or cure_all == True:
+                    statuses_list[ailment[0]].cure(target, index)
     for cure_item in target.cures_list:
         target.cures_list[cure_item] = False
 
@@ -173,6 +220,20 @@ antidote = Status("Antidote", "Makes you immune to common poisons", [["Lesser An
 cleanse = Status("Cleanse", "Makes you immune to almost any ailment", [["Cleanse", 0, "+"]], [["Cleanse", 0, "+"]], [], [], [])
 
 #########################################
+#                BUFFS                  #
+#########################################
+
+buff_fire_affinity = Buff("Fire Affinity", "Raises your Fire type damage multiplier", "Fire")
+buff_fire_resistance = Buff("Fire Resistance", "Raises your resistance to Fire type damage", "Fire")
+buff_Vitality = Buff("Vitality", "Raises your Vitality", "Vitality")
+buff_Strength = Buff("Strength", "Raises your Strength", "Strength")
+buff_Dexterity = Buff("Dexterity", "Raises your Dexterity", "Dexterity")
+buff_Intelligence = Buff("Intelligence", "Raises your Intelligence", "Intelligence")
+buff_Faith = Buff("Faith", "Raises your Faith", "Faith")
+buff_Speed = Buff("Speed", "Raises your Speed", "Speed")
+buff_XP_gain = Buff("XP Gain", "Raises your XP gain multiplier", "XP Gain")
+
+#########################################
 #           STANDARD EFFECTS            #
 #########################################
 
@@ -184,6 +245,9 @@ dexterity = Standard_Effect("Dexterity", 3, ["increased", "decreased"], "Dexteri
 intelligence = Standard_Effect("Intelligence", 3, ["increased", "decreased"], "Intelligence")
 faith = Standard_Effect("Faith", 3, ["increased", "decreased"], "Faith")
 speed = Standard_Effect("Speed", 3, ["increased", "decreased"], "Speed")
+affinity = Standard_Effect("Affinity", 3, ["increased", "decreased"], "Affinity")
+resistance = Standard_Effect("Resistance", 3, ["increased", "decreased"], "Resistance")
+xp_gain = Standard_Effect("XP Gain", 3, ["increased", "decreased"], "XP Gain")
 
 #########################################
 #                CURES                  #