+++ Changelog.md
@@ -1,8 +1,14 @@
 # Changelog
 
-## Version a0.2.4.1 (CURRENT)
+## Update a0.2.5 (6/23/23)
 - Changed the way status effects work on the back-end to be more modular; no difference to gameplay
 - Changed the way damage resistances work so they can now come in varying degrees of resistance as opposed to just weak, resistant, immune, or standard
+- Added encyclopedia system
+- Added armor system
+- Added buffs system
+- Added encyclopedia command
+- Added a few new armor pieces
+- Added a few new buffs
 
 ## Update a0.2.4 (6/22/23)
 - Added changelog :)